💚 Fix

+++ tests/two-envs/test_load_instance.py
@@ -1,3 +1,5 @@
+import os
+
 import pytest
 from lnhub_rest.routers.account import get_account_by_handle
 from lnhub_rest.routers.collaborator import delete_collaborator
@@ -32,37 +34,43 @@ def test_load_remote_instance():
 
 
 def test_load_after_revoked_access():
-    ln_setup.login("static-testuser1@lamin.ai", password="static-testuser1-password")
-    admin_token = ln_setup.settings.user.access_token
-    add_collaborator(
-        "static-testuser2",
-        "static-testuser1",
-        "static-testinstance1",
-        "write",
-        f"Bearer {admin_token}",
-    )
-    ln_setup.login("static-testuser2@lamin.ai", password="static-testuser2-password")
-    ln_setup.load("https://lamin.ai/static-testuser1/static-testinstance1", _test=True)
-    assert ln_setup.settings.instance.storage.root_as_str == "s3://lndb-setup-ci"
-    account = get_account_by_handle("static-testuser2")
-
-    delete_collaborator(
-        "static-testuser1",
-        "static-testinstance1",
-        account["id"],
-        f"Bearer {admin_token}",
-    )
-    with pytest.raises(RuntimeError) as error:
+    # can't currently test this on staging as I'm missing the accounts
+    if os.getenv("LAMIN_ENV") == "prod":
+        ln_setup.login(
+            "static-testuser1@lamin.ai", password="static-testuser1-password"
+        )
+        admin_token = ln_setup.settings.user.access_token
+        add_collaborator(
+            "static-testuser2",
+            "static-testuser1",
+            "static-testinstance1",
+            "write",
+            f"Bearer {admin_token}",
+        )
+        ln_setup.login(
+            "static-testuser2@lamin.ai", password="static-testuser2-password"
+        )
         ln_setup.load(
             "https://lamin.ai/static-testuser1/static-testinstance1", _test=True
         )
-
-    assert (
-        error.exconly()
-        == "RuntimeError: Remote instance static-testuser1/static-testinstance1 not"
-        " loadable from hub. The instance might have been deleted or you may have"
-        " lost access."
-    )
+        assert ln_setup.settings.instance.storage.root_as_str == "s3://lndb-setup-ci"
+        account = get_account_by_handle("static-testuser2")
+        delete_collaborator(
+            "static-testuser1",
+            "static-testinstance1",
+            account["id"],
+            f"Bearer {admin_token}",
+        )
+        with pytest.raises(RuntimeError) as error:
+            ln_setup.load(
+                "https://lamin.ai/static-testuser1/static-testinstance1", _test=True
+            )
+        assert (
+            error.exconly()
+            == "RuntimeError: Remote instance static-testuser1/static-testinstance1 not"
+            " loadable from hub. The instance might have been deleted or you may have"
+            " lost access."
+        )
 
 
 # def test_load_public_connection_string():
